+++ src/Parser.php
@@ -44,10 +44,6 @@ class Parser extends BaseObject
 
         $inlines = $grammar->getInlines();
         if ($inlines) {
-            // sort in reverse order to let more long items match first
-            // so /'$$' | '$'/ will find ['$$', '$'] in '$$$' and not ['$', '$', '$']
-            rsort($inlines, SORT_STRING);
-
             $my_lexer = $my_lexer->extend($inlines);
         }
         $this->lexer = $my_lexer;