+++ CHANGELOG.md
@@ -4,6 +4,7 @@ LR(0) parser Change Log
 1.3.2
 -----
 
+*   Add inline quoted tokens in grammar source.
 *   Add ability to omit some useless tokens from the resulting tree.
 
 1.3.0